Fix stx-libvirt container build failure on missing qemu
stx-libvirt container fails to build when it can't install package
'qemu'.

The package 'qemu' was recently upgraded to package qemu 7.3 from 5.2.
Qemu 7.2 no longer produces a dummy 'qemu' deb package.
Even in qemu 5.2 it was an empty deb package.

The container build instructions for stx-libvirt need to be updated,
removing the reference to package 'qemu'.

Partial-bug: 2051878
Change-Id: Ib01a220562dd44d386a4a9f4f94927a090986c1c
Signed-off-by: Scott Little <scott.little@windriver.com>