Zuul/Tox: Fix stx-virt-pylint node set
As per the reported launchpad [1], Zuul recently updated its default
nodeset from Ubuntu 22 to 24. Now, we need to set our jobs nodeset to
debian-bullsseye to properly execute it.
This was causing Zuul errors to any stx/virt change, but was not related
to the change content [2].

[1] https://bugs.launchpad.net/starlingx/+bug/2077748
[2] https://zuul.opendev.org/t/openstack/build/9f030173bfaa4bf68e43471df120736a

Test Plan:
PASS - stx-virt-pylint successfully executed

Partial-Bug: 2077748

Change-Id: I10e0b1e6f0d6d8b88914e779e05165150df4d42a
Signed-off-by: Thales Elero Cervi <thaleselero.cervi@windriver.com>